requirements/lints.txt
black==25.11.0
django-debug-toolbar==6.0.0
flake8==7.3.0
flake8-absolute-import==1.0.0.3
flake8-alphabetize==0.0.21
flake8-black==0.4.0
flake8-bugbear==25.11.29
flake8-clean-block==0.1.2
flake8-commas==4.0.0
flake8-dunder-all==0.5.0
flake8-expression-complexity==0.0.11
flake8-implicit-str-concat==0.6.0
flake8-import-order==0.19.2
flake8-quotes==3.4.0
flake8-return==1.2.0
pep8-naming==0.15.1